WIP: Rewrite Alarmserver on top of tornado.
