+++ .gitignore
@@ -2,3 +2,4 @@
 
 alarmserver.cfg
 .idea
+output.log
