+++ alarmserver/config.py
@@ -1,0 +1,66 @@
+import ConfigParser
+import logger
+
+MAXPARTITIONS=16
+MAXZONES=128
+MAXALARMUSERS=47
+
+class config():
+    def __init__(self, configfile):
+        self._config = ConfigParser.ConfigParser()
+        self._config.read(configfile)
+
+        self.LOGURLREQUESTS = self.read_config_var('alarmserver', 'logurlrequests', True, 'bool')
+        self.HTTPSPORT = self.read_config_var('alarmserver', 'httpsport', 8111, 'int')
+        self.CERTFILE = self.read_config_var('alarmserver', 'certfile', 'server.crt', 'str')
+        self.KEYFILE = self.read_config_var('alarmserver', 'keyfile', 'server.key', 'str')
+        self.MAXEVENTS = self.read_config_var('alarmserver', 'maxevents', 10, 'int')
+        self.MAXALLEVENTS = self.read_config_var('alarmserver', 'maxallevents', 100, 'int')
+        self.ENVISALINKHOST = self.read_config_var('envisalink', 'host', 'envisalink', 'str')
+        self.ENVISALINKPORT = self.read_config_var('envisalink', 'port', 4025, 'int')
+        self.ENVISALINKPASS = self.read_config_var('envisalink', 'pass', 'user', 'str')
+        self.ENABLEPROXY = self.read_config_var('envisalink', 'enableproxy', True, 'bool')
+        self.ENVISALINKPROXYPORT = self.read_config_var('envisalink', 'proxyport', self.ENVISALINKPORT, 'int')
+        self.ENVISALINKPROXYPASS = self.read_config_var('envisalink', 'proxypass', self.ENVISALINKPASS, 'str')
+        self.PUSHOVER_ENABLE = self.read_config_var('pushover', 'enable', False, 'bool')
+        self.PUSHOVER_USERTOKEN = self.read_config_var('pushover', 'enable', False, 'bool')
+        self.ALARMCODE = self.read_config_var('envisalink', 'alarmcode', 1111, 'int')
+        self.EVENTTIMEAGO = self.read_config_var('alarmserver', 'eventtimeago', True, 'bool')
+        self.LOGFILE = self.read_config_var('alarmserver', 'logfile', '', 'str')
+        if self.LOGFILE == '':
+            self.LOGTOFILE = False
+        else:
+            self.LOGTOFILE = True
+
+        self.PARTITIONNAMES={}
+        for i in range(1, MAXPARTITIONS+1):
+            partition = self.read_config_var('alarmserver', 'partition'+str(i), False, 'str', True)
+            if partition: self.PARTITIONNAMES[i] = partition
+
+        self.ZONENAMES={}
+        for i in range(1, MAXZONES+1):
+            zone = self.read_config_var('alarmserver', 'zone'+str(i), False, 'str', True)
+            if zone: self.ZONENAMES[i] = zone
+
+        self.ALARMUSERNAMES={}
+        for i in range(1, MAXALARMUSERS+1):
+            user = self.read_config_var('alarmserver', 'user'+str(i), False, 'str', True)
+            if user: self.ALARMUSERNAMES[i] = user
+
+        if self.PUSHOVER_USERTOKEN == False and self.PUSHOVER_ENABLE == True: self.PUSHOVER_ENABLE = False
+
+    def defaulting(self, section, variable, default, quiet = False):
+        if quiet == False:
+            logger.debug('Config option '+ str(variable) + ' not set in ['+str(section)+'] defaulting to: \''+str(default)+'\'')
+
+    def read_config_var(self, section, variable, default, type = 'str', quiet = False):
+        try:
+            if type == 'str':
+                return self._config.get(section,variable)
+            elif type == 'bool':
+                return self._config.getboolean(section,variable)
+            elif type == 'int':
+                return int(self._config.get(section,variable))
+        except (ConfigParser.NoSectionError, ConfigParser.NoOptionError):
+            self.defaulting(section, variable, default, quiet)
+            return default
